Lower memory usage in swpm_members prepare_items()
prepare_items() was retrieving every field in the join of swpm_members and swpm_membership, but only needed a few fields. This is a heavy query being run twice and can exceed available memory when there are many members and category/page protections. Change to only retrieve a count in the first query, and only the required columns in the second limited/paginated query.